Add view for creating a post on the web page
Add a view for creating a blog post on the webpage. Connect the urls to the view. Add the get_absolute_url to the Post model in models.py. Import reverse into the models.py file. Add a new template for the create post. Add basic content to the html. Add a link to the base.html to gain access to it from the webpage